updating note response types for chain endpoints
