+++ ironfish/src/rpc/routes/chain/getTransactionStream.ts
@@ -11,18 +11,9 @@ import { CurrencyUtils } from '../../../utils'
 import { PromiseUtils } from '../../../utils/promise'
 import { isValidIncomingViewKey } from '../../../wallet/validator'
 import { ValidationError } from '../../adapters/errors'
+import { RpcNote, RpcNoteSchema } from '../../types'
 import { ApiNamespace, routes } from '../router'
 
-interface Note {
-  assetId: string
-  /**
-   * @deprecated Please use getAsset endpoint to get this information
-   */
-  assetName: string
-  hash: string
-  value: string
-  memo: string
-}
 interface Mint {
   assetId: string
   /**
@@ -43,21 +34,10 @@ interface Burn {
 interface Transaction {
   hash: string
   isMinersFee: boolean
-  notes: Note[]
+  notes: RpcNote[]
   mints: Mint[]
   burns: Burn[]
 }
-
-const NoteSchema = yup
-  .object()
-  .shape({
-    assetId: yup.string().required(),
-    assetName: yup.string().required(),
-    hash: yup.string().required(),
-    value: yup.string().required(),
-    memo: yup.string().required(),
-  })
-  .required()
 
 const MintSchema = yup
   .object()
@@ -82,7 +62,7 @@ const TransactionSchema = yup
   .shape({
     hash: yup.string().required(),
     isMinersFee: yup.boolean().required(),
-    notes: yup.array().of(NoteSchema).required(),
+    notes: yup.array().of(RpcNoteSchema).required(),
     mints: yup.array().of(MintSchema).required(),
     burns: yup.array().of(BurnSchema).required(),
   })
@@ -163,7 +143,7 @@ routes.register<typeof GetTransactionStreamRequestSchema, GetTransactionStreamRe
       const transactions: Transaction[] = []
 
       for (const tx of block.transactions) {
-        const notes = new Array<Note>()
+        const notes = new Array<RpcNote>()
         const mints = new Array<Mint>()
         const burns = new Array<Burn>()
 
@@ -172,6 +152,7 @@ routes.register<typeof GetTransactionStreamRequestSchema, GetTransactionStreamRe
 
           if (decryptedNote) {
             const assetValue = await node.chain.getAssetById(decryptedNote.assetId())
+
             notes.push({
               value: CurrencyUtils.encode(decryptedNote.value()),
               memo: decryptedNote.memo(),
